Rename method module
