+++ src/multiview_stitcher/param_resolution/linear_two_pass.py
